requirements.txt
acres==0.3.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.5.0
attrs==25.3.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.13.3
blinker==1.4
bs4==0.0.2
certifi==2025.1.31
chardet==4.0.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.0.3
cloud-init==24.4.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
configparser==7.2.0
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
dill==0.4.0
distro==1.9.0
distro-info==1.1+ubuntu0.2
docling==2.30.0
docling-core==2.27.0
docling-ibm-models==3.4.1
docling-parse==4.0.1
dotenv==0.9.9
easyocr==1.7.2
et_xmlfile==2.0.0
etelemetry==0.3.1
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
fitz==0.0.1.dev2
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.30.2
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
importlib-metadata==4.6.4
importlib_resources==6.5.2
incremental==21.3.0
isodate==0.6.1
jeepney==0.7.1
Jinja2==3.0.3
jiter==0.9.0
jsonlines==3.1.0
jsonpatch==1.32
jsonpointer==2.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==23.5.0
latex2mathml==3.77.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
looseversion==1.3.0
lxml==5.3.2
markdown-it-py==3.0.0
marko==2.1.3
MarkupSafe==2.0.1
mdurl==0.1.2
more-itertools==8.10.0
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.18
netifaces==0.11.0
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.4
nipype==1.10.0
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
openai==1.72.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pexpect==4.8.0
pillow==11.1.0
pluggy==1.5.0
prov==2.0.1
ptyprocess==0.7.0
puremagic==1.28
pyasn1==0.4.8
pyasn1-modules==0.2.1
pyclipper==1.3.0.post6
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydot==3.0.4
Pygments==2.19.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pylatexenc==2.10
PyMuPDF==1.25.5
pyOpenSSL==21.0.0
pyparsing==3.2.3
pyparted==3.11.7
pypdfium2==4.30.1
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.13
python-apt==2.4.0+ubuntu4
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-docx==1.1.2
python-dotenv==1.1.0
python-magic==0.4.24
python-pptx==1.0.2
pytz==2022.1
pyxnat==1.6.3
PyYAML==5.4.1
rdflib==6.3.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
rtree==1.4.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
SecretStorage==3.3.1
selenium==4.31.0
semchunk==2.2.2
service-identity==18.1.0
shapely==2.1.0
shellingham==1.5.4
simplejson==3.20.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sos==4.7.2
soupsieve==2.6
ssh-import-id==5.11
starlette==0.46.1
sympy==1.13.1
systemd-python==234
tabulate==0.9.0
tifffile==2025.3.30
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traits==7.0.2
transformers==4.51.3
trio==0.29.0
trio-websocket==0.12.2
triton==3.2.0
Twisted==22.1.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.3.0
uvicorn==0.15.0
wadllib==1.3.6
WALinuxAgent==2.2.46
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.0.0
XlsxWriter==3.2.3
zipp==1.0.0
zope.interface==5.4.0
